Create Painter's partition problem
Given an array arr[] where each element denotes the length of a board, and an integer k representing the number of painters available. Each painter takes 1 unit of time to paint 1 unit length of a board.

Determine the minimum amount of time required to paint all the boards, under the constraint that each painter can paint only a contiguous sequence of boards (no skipping or splitting allowed).

Examples:

Input: arr[] = [5, 10, 30, 20, 15], k = 3
Output: 35
Explanation: The optimal allocation of boards among 3 painters is - 
Painter 1 → [5, 10] → time = 15
Painter 2 → [30] → time = 30
Painter 3 → [20, 15] → time = 35
Job will be done when all painters finish i.e. at time = max(15, 30, 35) = 35


Above is the problem statement which is a hard level problem , 


INTUTION : we can think how to apply binary search on this problem , but don't have the sorted array as input , and the main and interesting thing is that we don't havae to apply binary search on the array instead we have to apply binary search on a range where the answer can lie.

If we find one answer then we will not stop , instead we will store this answer and we will move to find the more smaller value untill the while loop ends


so after learning two problems from ma'am and i solved this problem as my own.....................................This is a quite interesting problem to solve