Create Rat in Maze - I
This a  very interesting problem whose solution can only be written if the person know recurson and backtraking quite well. I solved the whole problem as my own but one thing which i asked for others was making the cell visited during the recursive call.

By The Way, this was very intersting problem and i enjoyed solving this problem.